pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "integrao"
version = "0.1.1"
dependencies = [
    "snfpy",
    "networkx==3.0",
    "numpy>=1.24.1",
    "pandas>=1.3.5",
    "scikit-learn==1.3.2",
    "scipy==1.11.4",
    "umap-learn>=0.5.5",
    "torch-geometric>=2.3.0"
]
description = "The Python implementation of the IntegrAO."
authors = [
  { name="Shihao Ma", email="rex.ma@mail.utoronto.ca" },
]
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.urls]
Homepage = "https://github.com/bowang-lab/IntegrAO"
Issues = "https://github.com/bowang-lab/IntegrAO/issues"

